notify the read model when a http request has been writed or an item has been writed
