+++ Write Model/WriteModel.Infrastructure/Audit/CommandHandlers/AuditRequestHandler.cs
@@ -2,23 +2,22 @@
 using System.Threading.Tasks;
 using MediatR;
 using WriteModel.Domain.Audit.Aggregates;
-using WriteModel.Domain.Audit.Attributes;
 using WriteModel.Domain.Audit.Commands;
 using WriteModel.Domain.Audit.Configuration;
 using WriteModel.Domain.Audit.Services;
 using WriteModel.Domain.AuthentificationContext;
 using WriteModel.Infrastructure.Audit.DbContext;
 
-namespace WriteModel.Infrastructure.Audit.Handlers
+namespace WriteModel.Infrastructure.Audit.CommandHandlers
 {
-    public class AuditQueryHandler : IRequestHandler<CreateAuditQuery>
+    public class AuditRequestHandler : IRequestHandler<CreateAuditRequest>
     {
         private readonly IAuthentificationContext _authentificationContext;
         private readonly AuditDbContext _dbContext;
         private readonly IAuditSerializer _auditSerializer;
         private readonly AuditConfiguration _auditConfiguration;
 
-        public AuditQueryHandler(IAuthentificationContext authentificationContext, AuditDbContext dbContext, IAuditSerializer auditSerializer, AuditConfiguration auditConfiguration)
+        public AuditRequestHandler(IAuthentificationContext authentificationContext, AuditDbContext dbContext, IAuditSerializer auditSerializer, AuditConfiguration auditConfiguration)
         {
             _authentificationContext = authentificationContext;
             _dbContext = dbContext;
@@ -26,14 +25,14 @@ namespace WriteModel.Infrastructure.Audit.Handlers
             _auditConfiguration = auditConfiguration;
         }
 
-        public async Task<Unit> Handle(CreateAuditQuery query, CancellationToken cancellationToken)
+        public async Task<Unit> Handle(CreateAuditRequest command, CancellationToken cancellationToken)
         {
-            if (!(_auditConfiguration.AuditQueries && query.Query.GetType().ShouldAuditQuery() && _authentificationContext.IsValid()))
+            if (!(_auditConfiguration.AuditRequests && _authentificationContext.IsValid()))
                 return Unit.Value;
 
-            var auditQuery = AuditQuery.Create(query, _authentificationContext, _auditSerializer);
+            var auditRequest = AuditRequest.Create(command, _authentificationContext, _auditSerializer);
 
-            _dbContext.AuditQueries.Add(auditQuery);
+            _dbContext.AuditRequests.Add(auditRequest);
             await _dbContext.SaveChangesAsync();
 
             return Unit.Value;
